requirements.txt
tkinter
yt-dlp
sv-ttk
requests
Pillow
moviepy
